+++ algosdk/future/transaction.py
@@ -1,16 +1,16 @@
 import base64
 import binascii
 import warnings
-from collections import OrderedDict
+import msgpack
 from enum import IntEnum
 from typing import List, Union
+from collections import OrderedDict
 
-import msgpack
+from algosdk import account, constants, encoding, error, logic, transaction
+from algosdk.box_reference import BoxReference
+from algosdk.v2client import algod, models
 from nacl.exceptions import BadSignatureError
 from nacl.signing import SigningKey, VerifyKey
-
-from .. import account, constants, encoding, error, logic, transaction
-from ..v2client import algod, models
 
 
 class SuggestedParams:
@@ -1518,11 +1518,10 @@ class StateSchema:
 
     @staticmethod
     def undictify(d):
-        args = {
-            "num_uints": d["nui"] if "nui" in d else None,
-            "num_byte_slices": d["nbs"] if "nbs" in d else None,
-        }
-        return args
+        return StateSchema(
+            num_uints=d["nui"] if "nui" in d else None,
+            num_byte_slices=d["nbs"] if "nbs" in d else None,
+        )
 
     def __eq__(self, other):
         if not isinstance(other, StateSchema):
@@ -1583,6 +1582,7 @@ class ApplicationCallTxn(Transaction):
         foreign_apps (list[int], optional): list of other applications (identified by index) involved in call
         foreign_assets (list[int], optional): list of assets involved in call
         extra_pages (int, optional): additional program space for supporting larger programs.  A page is 1024 bytes.
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
 
     Attributes:
         sender (str)
@@ -1601,6 +1601,7 @@ class ApplicationCallTxn(Transaction):
         foreign_apps (list[int])
         foreign_assets (list[int])
         extra_pages (int)
+        boxes (list[(int, bytes)])
     """
 
     def __init__(
@@ -1621,6 +1622,7 @@ class ApplicationCallTxn(Transaction):
         lease=None,
         rekey_to=None,
         extra_pages=0,
+        boxes=None,
     ):
         Transaction.__init__(
             self, sender, sp, note, lease, constants.appcall_txn, rekey_to
@@ -1636,6 +1638,7 @@ class ApplicationCallTxn(Transaction):
         self.foreign_apps = self.int_list(foreign_apps)
         self.foreign_assets = self.int_list(foreign_assets)
         self.extra_pages = extra_pages
+        self.boxes = BoxReference.translate_box_references(boxes, self.foreign_apps, self.index)  # type: ignore
         if not sp.flat_fee:
             self.fee = max(
                 self.estimate_size() * self.fee, constants.min_txn_fee
@@ -1646,9 +1649,8 @@ class ApplicationCallTxn(Transaction):
         """Confirm the argument is a StateSchema, or false which is coerced to None"""
         if not schema or not schema.dictify():
             return None  # Coerce false/empty values to None, to help __eq__
-        assert isinstance(
-            schema, StateSchema
-        ), "{} is not a StateSchema".format(schema)
+        if not isinstance(schema, StateSchema):
+            raise TypeError("{} is not a StateSchema".format(schema))
         return schema
 
     @staticmethod
@@ -1656,28 +1658,16 @@ class ApplicationCallTxn(Transaction):
         """Confirm the argument is bytes-like, or false which is coerced to None"""
         if not teal:
             return None  # Coerce false values like "" to None, to help __eq__
-        assert isinstance(
-            teal, (bytes, bytearray)
-        ), "Program {} is not bytes".format(teal)
+        if not isinstance(teal, (bytes, bytearray)):
+            raise TypeError("Program {} is not bytes".format(teal))
         return teal
 
     @staticmethod
     def bytes_list(lst):
         """Confirm or coerce list elements to bytes. Return None for empty/false lst."""
-
-        def as_bytes(e):
-            if isinstance(e, (bytes, bytearray)):
-                return e
-            if isinstance(e, str):
-                return e.encode()
-            if isinstance(e, int):
-                # Uses 8 bytes, big endian to match TEAL's btoi
-                return e.to_bytes(8, "big")  # raises for negative or too big
-            assert False, "{} is not bytes, str, or int".format(e)
-
         if not lst:
             return None
-        return [as_bytes(elt) for elt in lst]
+        return [encoding.encode_as_bytes(elt) for elt in lst]
 
     @staticmethod
     def int_list(lst):
@@ -1712,6 +1702,8 @@ class ApplicationCallTxn(Transaction):
             d["apas"] = self.foreign_assets
         if self.extra_pages:
             d["apep"] = self.extra_pages
+        if self.boxes:
+            d["apbx"] = [box.dictify() for box in self.boxes]
 
         d.update(super(ApplicationCallTxn, self).dictify())
         od = OrderedDict(sorted(d.items()))
@@ -1723,10 +1715,10 @@ class ApplicationCallTxn(Transaction):
         args = {
             "index": d["apid"] if "apid" in d else None,
             "on_complete": d["apan"] if "apan" in d else None,
-            "local_schema": StateSchema(**StateSchema.undictify(d["apls"]))
+            "local_schema": StateSchema.undictify(d["apls"])
             if "apls" in d
             else None,
-            "global_schema": StateSchema(**StateSchema.undictify(d["apgs"]))
+            "global_schema": StateSchema.undictify(d["apgs"])
             if "apgs" in d
             else None,
             "approval_program": d["apap"] if "apap" in d else None,
@@ -1736,6 +1728,9 @@ class ApplicationCallTxn(Transaction):
             "foreign_apps": d["apfa"] if "apfa" in d else None,
             "foreign_assets": d["apas"] if "apas" in d else None,
             "extra_pages": d["apep"] if "apep" in d else 0,
+            "boxes": [BoxReference.undictify(box) for box in d["apbx"]]
+            if "apbx" in d
+            else None,
         }
         if args["accounts"]:
             args["accounts"] = [
@@ -1760,6 +1755,7 @@ class ApplicationCallTxn(Transaction):
             and self.foreign_apps == other.foreign_apps
             and self.foreign_assets == other.foreign_assets
             and self.extra_pages == other.extra_pages
+            and self.boxes == other.boxes
         )
 
 
@@ -1783,6 +1779,7 @@ class ApplicationCreateTxn(ApplicationCallTxn):
         lease(bytes, optional): transaction lease field
         rekey_to(str, optional): rekey-to field, see Transaction
         extra_pages(int, optional): provides extra program size
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
 
     Attributes:
         See ApplicationCallTxn
@@ -1805,6 +1802,7 @@ class ApplicationCreateTxn(ApplicationCallTxn):
         lease=None,
         rekey_to=None,
         extra_pages=0,
+        boxes=None,
     ):
         ApplicationCallTxn.__init__(
             self,
@@ -1824,6 +1822,7 @@ class ApplicationCreateTxn(ApplicationCallTxn):
             lease=lease,
             rekey_to=rekey_to,
             extra_pages=extra_pages,
+            boxes=boxes,
         )
 
 
@@ -1844,6 +1843,8 @@ class ApplicationUpdateTxn(ApplicationCallTxn):
         note(bytes, optional): transaction note field
         lease(bytes, optional): transaction lease field
         rekey_to(str, optional): rekey-to field, see Transaction
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
+
 
     Attributes:
         See ApplicationCallTxn
@@ -1863,6 +1864,7 @@ class ApplicationUpdateTxn(ApplicationCallTxn):
         note=None,
         lease=None,
         rekey_to=None,
+        boxes=None,
     ):
         ApplicationCallTxn.__init__(
             self,
@@ -1879,6 +1881,7 @@ class ApplicationUpdateTxn(ApplicationCallTxn):
             note=note,
             lease=lease,
             rekey_to=rekey_to,
+            boxes=boxes,
         )
 
 
@@ -1897,6 +1900,7 @@ class ApplicationDeleteTxn(ApplicationCallTxn):
         note(bytes, optional): transaction note field
         lease(bytes, optional): transaction lease field
         rekey_to(str, optional): rekey-to field, see Transaction
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
 
     Attributes:
         See ApplicationCallTxn
@@ -1914,6 +1918,7 @@ class ApplicationDeleteTxn(ApplicationCallTxn):
         note=None,
         lease=None,
         rekey_to=None,
+        boxes=None,
     ):
         ApplicationCallTxn.__init__(
             self,
@@ -1928,6 +1933,7 @@ class ApplicationDeleteTxn(ApplicationCallTxn):
             note=note,
             lease=lease,
             rekey_to=rekey_to,
+            boxes=boxes,
         )
 
 
@@ -1946,6 +1952,7 @@ class ApplicationOptInTxn(ApplicationCallTxn):
         note(bytes, optional): transaction note field
         lease(bytes, optional): transaction lease field
         rekey_to(str, optional): rekey-to field, see Transaction
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
 
     Attributes:
         See ApplicationCallTxn
@@ -1963,6 +1970,7 @@ class ApplicationOptInTxn(ApplicationCallTxn):
         note=None,
         lease=None,
         rekey_to=None,
+        boxes=None,
     ):
         ApplicationCallTxn.__init__(
             self,
@@ -1977,6 +1985,7 @@ class ApplicationOptInTxn(ApplicationCallTxn):
             note=note,
             lease=lease,
             rekey_to=rekey_to,
+            boxes=boxes,
         )
 
 
@@ -1995,6 +2004,7 @@ class ApplicationCloseOutTxn(ApplicationCallTxn):
         note(bytes, optional): transaction note field
         lease(bytes, optional): transaction lease field
         rekey_to(str, optional): rekey-to field, see Transaction
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
 
     Attributes:
         See ApplicationCallTxn
@@ -2012,6 +2022,7 @@ class ApplicationCloseOutTxn(ApplicationCallTxn):
         note=None,
         lease=None,
         rekey_to=None,
+        boxes=None,
     ):
         ApplicationCallTxn.__init__(
             self,
@@ -2026,6 +2037,7 @@ class ApplicationCloseOutTxn(ApplicationCallTxn):
             note=note,
             lease=lease,
             rekey_to=rekey_to,
+            boxes=boxes,
         )
 
 
@@ -2044,6 +2056,7 @@ class ApplicationClearStateTxn(ApplicationCallTxn):
         note(bytes, optional): transaction note field
         lease(bytes, optional): transaction lease field
         rekey_to(str, optional): rekey-to field, see Transaction
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
 
     Attributes:
         See ApplicationCallTxn
@@ -2061,6 +2074,7 @@ class ApplicationClearStateTxn(ApplicationCallTxn):
         note=None,
         lease=None,
         rekey_to=None,
+        boxes=None,
     ):
         ApplicationCallTxn.__init__(
             self,
@@ -2075,6 +2089,7 @@ class ApplicationClearStateTxn(ApplicationCallTxn):
             note=note,
             lease=lease,
             rekey_to=rekey_to,
+            boxes=boxes,
         )
 
 
@@ -2094,6 +2109,7 @@ class ApplicationNoOpTxn(ApplicationCallTxn):
         note(bytes, optional): transaction note field
         lease(bytes, optional): transaction lease field
         rekey_to(str, optional): rekey-to field, see Transaction
+        boxes(list[(int, bytes)], optional): list of tuples specifying app id and key for boxes the app may access
 
     Attributes:
         See ApplicationCallTxn
@@ -2111,6 +2127,7 @@ class ApplicationNoOpTxn(ApplicationCallTxn):
         note=None,
         lease=None,
         rekey_to=None,
+        boxes=None,
     ):
         ApplicationCallTxn.__init__(
             self,
@@ -2125,6 +2142,7 @@ class ApplicationNoOpTxn(ApplicationCallTxn):
             note=note,
             lease=lease,
             rekey_to=rekey_to,
+            boxes=boxes,
         )